feat: naming and tidying
